GUI Task: Classes sorted to packages, code lines optimozed, code convention used
