+++ .gitignore
@@ -18,10 +18,9 @@
 *.zip
 *.tar.gz
 *.rar
-
 # IntellijIdea files
 *.iml
-.idea
+*.idea
 out
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
