--- src/System/Circle.java
@@ -1,10 +1,0 @@
-package System;
-
-
-public class Circle extends Ellipse {
-
-	public Circle(){
-
-	}
-
-}//end Circle
