--- src/System/Ellipse.java
@@ -1,27 +1,0 @@
-package System;
-
-
-import java.awt.*;
-
-class Ellipse extends Figure2D {
-
-	private Point firstPoint;
-	private Point secondPoint;
-
-	public Ellipse(){
-
-	}
-
-	public void draw(){
-
-	}
-
-	/**
-	 * 
-	 * @param p2
-	 * @param p1
-	 */
-	public void setcenter(Point p2, Point p1){
-
-	}
-}//end Ellipse
