--- src/System/IsoscelesTriangle.java
@@ -1,9 +1,0 @@
-package System;
-
-public class IsoscelesTriangle extends SymmetricFigure {
-
-	public IsoscelesTriangle(){
-
-	}
-
-}//end IsoscelesTriangle
