--- src/System/Line.java
@@ -1,12 +1,0 @@
-package System;
-
-public class Line extends Ray {
-
-	public Line(){
-
-	}
-
-	public void draw(){
-
-	}
-}//end Line
