--- src/System/Ray.java
@@ -1,12 +1,0 @@
-package System;
-
-public class Ray extends Segment {
-
-	public Ray(){
-
-	}
-
-	public void draw(){
-
-	}
-}//end Ray
