--- src/System/Rectangle.java
@@ -1,9 +1,0 @@
-package System;
-
-public class Rectangle extends SymmetricFigure {
-
-	public Rectangle(){
-
-	}
-
-}//end Rectangle
