--- src/System/Rhombus.java
@@ -1,9 +1,0 @@
-package System;
-
-public class Rhombus extends SymmetricFigure {
-
-	public Rhombus(){
-
-	}
-
-}//end Rhombus
