--- src/System/Segment.java
@@ -1,29 +1,0 @@
-package System;
-
-
-import java.awt.*;
-
-public class Segment extends Figure1D {
-
-	public PolygonalChain m_PolygonalChain;
-
-	public Segment(){
-
-	}
-
-	public void draw(){
-
-	}
-
-	/**
-	 * 
-	 * @param p1
-	 */
-	public void move(Point p1){
-
-	}
-
-	public void setcenter(){
-
-	}
-}//end Segment
