+++ src/System/figures/figures1d/PolygonalChain.java
@@ -1,0 +1,41 @@
+package System.figures.figures1d;
+
+import System.figures.figures1d.segments.Segment;
+
+import java.awt.*;
+import java.util.ArrayList;
+
+public class PolygonalChain {
+    private ArrayList<Segment> segments;
+
+    public PolygonalChain() {
+
+    }
+
+    public void draw() {
+
+    }
+
+    public ArrayList<Segment> getSegments() {
+        return segments;
+    }
+
+    /**
+     * @param p2
+     * @param p1
+     */
+    public void move(Point p2, Point p1) {
+
+    }
+
+    public void setCenter() {
+
+    }
+
+    /**
+     * @param newVal
+     */
+    public void setSegments(ArrayList<Segment> newVal) {
+        segments = newVal;
+    }
+}//end PolygonalChain
