+++ src/System/figures/figures1d/segments/Line.java
@@ -1,0 +1,13 @@
+package System.figures.figures1d.segments;
+
+import System.figures.figures1d.segments.Ray;
+
+public class Line extends Ray {
+    public Line() {
+
+    }
+
+    public void draw() {
+
+    }
+}//end Line
