+++ src/System/figures/figures1d/segments/Ray.java
@@ -1,0 +1,13 @@
+package System.figures.figures1d.segments;
+
+import System.figures.figures1d.segments.Segment;
+
+public class Ray extends Segment {
+    public Ray() {
+
+    }
+
+    public void draw() {
+
+    }
+}//end Ray
