+++ src/System/figures/figures1d/segments/Segment.java
@@ -1,0 +1,30 @@
+package System.figures.figures1d.segments;
+
+
+import System.figures.figures1d.Figure1D;
+import System.figures.figures1d.PolygonalChain;
+
+import java.awt.*;
+
+public class Segment extends Figure1D {
+    public PolygonalChain polygonalChain;
+
+    public Segment() {
+
+    }
+
+    public void draw() {
+
+    }
+
+    /**
+     * @param p1
+     */
+    public void move(Point p1) {
+
+    }
+
+    public void setCenter() {
+
+    }
+}//end Segment
