+++ src/System/figures/figures2d/ellipse/Circle.java
@@ -1,0 +1,10 @@
+package System.figures.figures2d.ellipse;
+
+
+public class Circle extends Ellipse {
+
+	public Circle(){
+
+	}
+
+}//end Circle
