+++ src/System/figures/figures2d/ellipse/Ellipse.java
@@ -1,0 +1,27 @@
+package System.figures.figures2d.ellipse;
+
+
+import System.figures.figures2d.Figure2D;
+
+import java.awt.*;
+
+class Ellipse extends Figure2D {
+    private Point firstPoint;
+    private Point secondPoint;
+
+    public Ellipse() {
+
+    }
+
+    public void draw() {
+
+    }
+
+    /**
+     * @param p2
+     * @param p1
+     */
+    public void setCenter(Point p2, Point p1) {
+
+    }
+}//end Ellipse
