+++ src/System/figures/figures2d/polygons/symmetric/IsoscelesTriangle.java
@@ -1,0 +1,8 @@
+package System.figures.figures2d.polygons.symmetric;
+
+public class IsoscelesTriangle extends SymmetricFigure {
+    public IsoscelesTriangle() {
+
+    }
+
+}//end IsoscelesTriangle
