+++ src/System/figures/figures2d/polygons/symmetric/Rectangle.java
@@ -1,0 +1,8 @@
+package System.figures.figures2d.polygons.symmetric;
+
+public class Rectangle extends SymmetricFigure {
+    public Rectangle() {
+
+    }
+
+}//end Rectangle
