+++ src/System/figures/figures2d/polygons/symmetric/Rhombus.java
@@ -1,0 +1,8 @@
+package System.figures.figures2d.polygons.symmetric;
+
+public class Rhombus extends SymmetricFigure {
+    public Rhombus() {
+
+    }
+
+}//end Rhombus
